moved to a somehow cleaner handling of api changes
Instead of implementing the functions directly, the vmod now relies on
a simpler implementation. The vmod glue is handled separately. In the
process the useless init function disappeared...

And yes it was done because of another API change in the trunk. I'm not
blaming anyone here, Varnish 4 is still under development ;)