refactored (reg)filter and clean functions
OOPs, I did it again. In order to tackle code duplication, the
implementation is sort-of object-oriented. Not only did duplicate code
go away, but it also became abstract enough to become the `clean'
function backbone.

Another side effect of this refactoring is that all three function are
now enhanced with better garbage cleaning, hence the new tests.