Build an actual list for the filter
Instead of dragging the va_list everywhere, let's first build a proper
list. It will allow the reuse of a filter list created outside of a
transaction. If it's part of a transaction, then nodes are cheaply
allocated from a workspace and don't need any cleanup other than a reset
of the workspace.