Save a workspace allocation in the sort function
The sort function would do a fixed-size allocation of the result URL
(because we know its maximum size) and a variable-size allocation of
a query-parameters table.

That would make use of:

- WS_Alloc
- WS_Snapshot
- WS_Reserve
- WS_Reset
- WS_Release

By not using `WS_Alloc` it is no longer needed to call `WS_Reset` since
a release will be enough. Also if the sorted URL is smaller (it cleans
on top of sorting) then we can save a marginal amount of one byte per
spurious &.