Avoid overflow-prone workspace allocations
Instead, do a safe unbounded reservation and release the excess. As a
result `qs_apply` now performs a single reservation for both the result
and tracking of params. This also means that there's no longer a need
for snapshots, getting rid of the fugly dirty hacks for Varnish 5.1+
compatibility...

The net result is that vmod-querystring now truly relies strictly on
VRT and advertises it in the VCC descriptor.