Don't jail varnishd in the test suite
When running as root, provided that users and groups exist varnishd uses
the unix jail by default. This fails the test suite when running sbuild
on Travis CI because the privileges drop no longer allows to reach the
module's shared object.

The problem is that the travis should be part of the sbuild group to run
make sbuild, but sbuild is installed as part of the build job so there's
no way at that point to log out and in again to gain sbuild privileges.

So sbuild is run via sudo since it's (obviously) configured as non
interactive in a CI context, but then the jail kicks in... One solution
could have been to redifine the ${varnishd} macro as varnishd -j none
directly in the makefile but it no longer exists in 6.0 and later.

Piling up one hack over the next ain't that bad if it's the last straw.