Conditional support for VCL_REGEX
In order to support both the STRING or REGEX argument in the regex
method, to target Varnish versions with or without the VCL_REGEX type,
one would normally use the autoconf macro VARNISH_VMODS_GENERATED.

Until it is available with all supported Varnish versions, when 6.0 LTS
goes EOL, it is not portable to generate the VCC file at build time. It
can instead be generated at configure time, the VARNISH_VMODS_GENERATED
macro is emulated in the top makefile until then.

This should be enough to support both 6.0 LTS with the original VRE API,
and 6.6.0 or later relying entirely on VRT. The 7.0.0 revised VRE API is
no longer a problem since this module no longer uses it.

Closes #49
Fixes #51